chore: pageディレクトリを使用するように変更
