+++ src/page/page.tsx
@@ -1,0 +1,115 @@
+import { format } from '@formkit/tempo';
+import Image from 'next/image';
+
+export default function Home() {
+  return (
+    <main className='flex min-h-screen flex-col items-center justify-between p-24'>
+      <div className='z-10 w-full max-w-5xl items-center justify-between font-mono text-sm lg:flex'>
+        <p className='fixed top-0 left-0 flex w-full justify-center border-gray-300 border-b bg-gradient-to-b from-zinc-200 pt-8 pb-6 backdrop-blur-2xl lg:static lg:w-auto lg:rounded-xl lg:border dark:border-neutral-800 dark:bg-zinc-800/30 lg:bg-gray-200 lg:dark:bg-zinc-800/30 dark:from-inherit lg:p-4'>
+          Get started by editing&nbsp;
+          <code className='font-bold font-mono'>src/app/page.tsx</code>
+        </p>
+        <div className='fixed bottom-0 left-0 flex h-48 w-full items-end justify-center bg-gradient-to-t from-white via-white lg:static lg:size-auto lg:bg-none dark:from-black dark:via-black'>
+          <a
+            className='pointer-events-none flex place-items-center gap-2 p-8 lg:pointer-events-auto lg:p-0'
+            href='https://vercel.com?utm_source=create-next-app&utm_medium=appdir-template&utm_campaign=create-next-app'
+            target='_blank'
+            rel='noopener noreferrer'
+          >
+            By{' '}
+            <Image
+              src='/vercel.svg'
+              alt='Vercel Logo'
+              className='dark:invert'
+              width={100}
+              height={24}
+              priority
+            />
+          </a>
+        </div>
+      </div>
+
+      <div className='before:-translate-x-1/2 after:-z-20 relative z-[-1] flex flex-col place-items-center gap-8 after:absolute before:absolute after:h-[180px] before:h-[300px] before:lg:h-[360px] after:w-full before:w-full sm:after:w-[240px] sm:before:w-[480px] after:translate-x-1/3 before:rounded-full after:bg-gradient-conic before:bg-gradient-radial before:dark:bg-gradient-to-br after:dark:from-sky-900 after:from-sky-200 before:dark:from-transparent before:from-white after:dark:via-[#0141ff] after:via-blue-200 before:dark:to-blue-700 before:to-transparent after:dark:opacity-40 before:dark:opacity-10 after:blur-2xl before:blur-2xl after:content-[""] before:content-[""]'>
+        <Image
+          className='relative dark:drop-shadow-[0_0_0.3rem_#ffffff70] dark:invert'
+          src='/next.svg'
+          alt='Next.js Logo'
+          width={180}
+          height={37}
+          priority
+        />
+        <p>{format(new Date(), 'MM/DD HH:mm')}</p>
+      </div>
+
+      <div className='mb-32 grid text-center lg:mb-0 lg:w-full lg:max-w-5xl lg:grid-cols-4 lg:text-left'>
+        <a
+          href='https://nextjs.org/docs?utm_source=create-next-app&utm_medium=appdir-template&utm_campaign=create-next-app'
+          className='group rounded-lg border border-transparent px-5 py-4 transition-colors hover:border-gray-300 hover:dark:border-neutral-700 hover:bg-gray-100 hover:dark:bg-neutral-800/30'
+          target='_blank'
+          rel='noopener noreferrer'
+        >
+          <h2 className='mb-3 font-semibold text-2xl'>
+            Docs{' '}
+            <span className='inline-block transition-transform group-hover:translate-x-1 motion-reduce:transform-none'>
+              -&gt;
+            </span>
+          </h2>
+          <p className='m-0 max-w-[30ch] text-sm opacity-50'>
+            Find in-depth information about Next.js features and API.
+          </p>
+        </a>
+
+        <a
+          href='https://nextjs.org/learn?utm_source=create-next-app&utm_medium=appdir-template-tw&utm_campaign=create-next-app'
+          className='group rounded-lg border border-transparent px-5 py-4 transition-colors hover:border-gray-300 hover:dark:border-neutral-700 hover:bg-gray-100 hover:dark:bg-neutral-800/30'
+          target='_blank'
+          rel='noopener noreferrer'
+        >
+          <h2 className='mb-3 font-semibold text-2xl'>
+            Learn{' '}
+            <span className='inline-block transition-transform group-hover:translate-x-1 motion-reduce:transform-none'>
+              -&gt;
+            </span>
+          </h2>
+          <p className='m-0 max-w-[30ch] text-sm opacity-50'>
+            Learn about Next.js in an interactive course with&nbsp;quizzes!
+          </p>
+        </a>
+
+        <a
+          href='https://vercel.com/templates?framework=next.js&utm_source=create-next-app&utm_medium=appdir-template&utm_campaign=create-next-app'
+          className='group rounded-lg border border-transparent px-5 py-4 transition-colors hover:border-gray-300 hover:dark:border-neutral-700 hover:bg-gray-100 hover:dark:bg-neutral-800/30'
+          target='_blank'
+          rel='noopener noreferrer'
+        >
+          <h2 className='mb-3 font-semibold text-2xl'>
+            Templates{' '}
+            <span className='inline-block transition-transform group-hover:translate-x-1 motion-reduce:transform-none'>
+              -&gt;
+            </span>
+          </h2>
+          <p className='m-0 max-w-[30ch] text-sm opacity-50'>
+            Explore starter templates for Next.js.
+          </p>
+        </a>
+
+        <a
+          href='https://vercel.com/new?utm_source=create-next-app&utm_medium=appdir-template&utm_campaign=create-next-app'
+          className='group rounded-lg border border-transparent px-5 py-4 transition-colors hover:border-gray-300 hover:dark:border-neutral-700 hover:bg-gray-100 hover:dark:bg-neutral-800/30'
+          target='_blank'
+          rel='noopener noreferrer'
+        >
+          <h2 className='mb-3 font-semibold text-2xl'>
+            Deploy{' '}
+            <span className='inline-block transition-transform group-hover:translate-x-1 motion-reduce:transform-none'>
+              -&gt;
+            </span>
+          </h2>
+          <p className='m-0 max-w-[30ch] text-balance text-sm opacity-50'>
+            Instantly deploy your Next.js site to a shareable URL with Vercel.
+          </p>
+        </a>
+      </div>
+    </main>
+  );
+}
